+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented in this file.
 
-### [5.0.1] 2021-03-20
+### [5.0.1] 2021-03-19
 
 #### Changed
 
@@ -12,7 +12,7 @@ All notable changes to this project are documented in this file.
 
 #### Added
 
-- univesal node: player.get.buttonlockstate, player.set.buttonlockstate
+- universal node: player.get.buttonlockstate, player.set.buttonlockstate
 
 - universal node: group.play.httpstream may use msg.info, msg.artUri to display title, use cover